Added exception class and handling, input validation, bug fixing, changed operational data type from int to double.

+++ AddOperatorCommand.h
@@ -8,7 +8,7 @@
 #include "Command.h"
 
 class AddOperatorCommand: public Command {
-    virtual void execute(std::stack <int> &);
+    virtual void execute(std::stack <double> &);
 };
